requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.6.4
pydantic-settings==2.1.0

# LLM & AI
openai==1.12.0
anthropic==0.25.0
litellm==1.32.0
langchain==0.1.10
langchain-openai==0.0.8
langchain-anthropic==0.1.8

# Code Execution
subprocess32==3.5.4
timeout-decorator==1.1.1
RestrictedPython==6.2

# Computer Control & Vision
opencv-python==4.8.1.78
pyautogui==0.9.54
pillow==10.1.0
numpy==1.24.3
pyscreenshot==3.1

# Audio & Voice
openai-whisper==20240314
librosa==0.10.0
soundfile==0.12.1

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
redis==5.0.1

# Utilities
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
pyyaml==6.0.1
rich==13.7.0
typer==0.9.0
click==8.1.7
psutil==5.9.6
platformdirs==4.2.0

# Async & Concurrency
asyncio==3.4.3
aiofiles==23.2.1

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.0
isort==5.13.2
flake8==6.1.0

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.39.1

# Security
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Diagram & Visualization
graphviz==0.20.1
mermaid==0.1.0
plotly==5.18.0
matplotlib==3.8.2
